requirements.txt
# PyTorch & CUDA 11.8 Support
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.2.1+cu118
torchvision==0.17.1+cu118
torchaudio==2.2.1+cu118

# Model Architecture & DL Utils
timm==1.0.15
ptflops==0.7.3
segment-anything==1.0
clip==1.0

# Computer Vision & Image Processing
opencv-python==4.11.0.86
Pillow==10.2.0

# Data Science & Visualization
numpy==1.24.1
scikit-learn==1.3.2
matplotlib==3.7.5
scipy==1.10.1

# Utilities
tqdm==4.67.1
PyYAML==6.0.2